refactor: relocate test fixtures and a stray helper
- engine/mod.rs was 75% test scaffolding and result/stats.rs held the
  fixtures four sibling modules reach into, so both get their own file.
- ensure_events_loaded lived in charts.rs but every caller is in
  corporate.rs.
- Restore a comment dropped from test_interval_duration_secs and rewrap
  the rolling.rs drawdown link as a reference definition.
